Move program to monitor memory usage to other package
Previously, code to run program a number of times and analyze memory usage was
in a package called "Test". This wasn't suited to what it was, and also wasn't
appropriate for further additions to performance monitoring.
This commit fixes it by creating a new package for performance analysis and
moving the relevant code to that package.